requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.1
pyarrow>=14.0.0

# Web scraping
playwright>=1.40.0
scrapy>=2.11.0
lxml>=4.9.0

# Configuration and logging
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=23.2.0

# Data validation
pydantic>=2.5.0
cerberus>=1.3.5

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Monitoring and performance
psutil>=5.9.0
memory-profiler>=0.61.0
